requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
pandas==2.1.4
openpyxl==3.1.2
pyxlsb==1.0.10
# Downgraded to stable combo known to work with Kaleido
plotly==5.17.0
kaleido==0.2.1
python-docx==1.1.0
Pillow==10.1.0
xlsxwriter==3.1.9
Werkzeug==3.0.1
gunicorn==21.2.0
